+++ scripts/build.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env sh
 
-go build -o sway-yast cmd/sway-yast/main.go
+go build -o sway-yasm cmd/sway-yasm/main.go