+++ CHANGELOG.md
@@ -1,12 +1,16 @@
-## dev
+## v0.2.0
 
+- refac: rearrange under pressure
+- feat: add clipboard manager
 - feat: add support for user command files
-- fix: correct app ID on startup
-- feat(cmd): add win-to-space
+- fix: correct app IDs on startup
+- feat(usr-cmd): add titlebar-toggle
 - feat(usr-cmd): add resize-toggle
 - feat(usr-cmd): add arrange
+- feat(cmd): add win-to-space
 
 ## v0.1.0
+
 - mouse follows focus
 - pick space
 - pick win
@@ -14,4 +18,5 @@
 - live config changes
 
 ## v0.0.1
+
 - initial release